+++ RELEASE_NOTES.md
@@ -1,5 +1,16 @@
 # AndroidSidebar - RELEASE NOTES
 
+## Version 1.3.0 (Jan. 3rd 2016)
+
+A feature release, which introduces the following features:
+
+- Added `ensureFileIsNoDirectory`-method to class `Condition`.
+- Added `getDeviceType`-method to class `DisplayUtil`.
+- Added the utility class `AppUtil`, which provides methods, which allow to start system apps.
+- Added the utility class `PermissionUtil`, which provides methods, which allow to handle permissions.
+- Removed the annotation `@VisibleForTesting` as the annotation support library provides an identical one.
+- As the library does now depend on the v4 support library, the minimum SDK version has been increased to 4.
+
 ## Version 1.2.4 (Dec. 27th 2015)
 
 A minor release, which introduces the following changes: